third_party/blink/web_tests/external/wpt/service-workers/service-worker/fetch-event.https-expected.txt

This is a testharness.js-based test.
[FAIL] FetchEvent#request.isReloadNavigation is true (location.reload())
  assert_equals: expected "method = GET, isReloadNavigation = false" but got "method = GET, isReloadNavigation = undefined"
[FAIL] FetchEvent#request.isReloadNavigation is true (history.go(0))
  assert_equals: expected "method = GET, isReloadNavigation = false" but got "method = GET, isReloadNavigation = undefined"
[FAIL] FetchEvent#request.isReloadNavigation is true (POST + location.reload())
  assert_equals: expected "method = GET, isReloadNavigation = false" but got "method = GET, isReloadNavigation = undefined"
[FAIL] FetchEvent#request.isReloadNavigation is true (with history traversal)
  assert_equals: expected "method = GET, isReloadNavigation = false" but got "method = GET, isReloadNavigation = undefined"
Harness: the test ran to completion.
